// src/app/shared/models/subprocess-info.model.ts
export class SubProcessTypeInfoDto {
  subProcessInfoId: number;
  processInfoId: number;
  subProcessTypeId: number;
  lengthOfCut?: number;
  formLength?: number;
  formHeight?: number;
  hlFactor?: number;
  bendingLineLength?: number;
  shoulderWidth?: number;
  noOfBends?: number;
  formPerimeter?: number;
  recommendTonnage?: number;
  diameterOfDrill?: number;
  depthOfDrill?: number;
  noOfPasses?: number;
  noOfHoles?: number;
  noOfMultiples?: number;
  cuttingTime?: number;
  cycleTime?: number;
  workpieceInitialDia?: number;
  workpieceFinalDia?: number;
  workpieceOuterDia?: number;
  workpieceInnerDia?: number;
  partInitialDia?: number;
  finalGrooveDia?: number;
  widthOfCut?: number;
  totalDepOfCut?: number;
  drillDiameter?: number;
  totalNoOfPasses?: number;
  wheelDiameter?: number;
  wheelWidth?: number;
  cutterDia?: number;
  noOfToolInserts?: number;
  fluteLength?: number;
  noOfFlutes?: number;
  noOfTeeth?: number;
  pitchDiameter?: number;
  outerDiameter?: number;
  rootDiameter?: number;
  pressureAngle?: number;
  helixAngle?: number;
  finishCutMatAllowance?: number;
  hobDiameter?: number;
  noOfStarts?: number;
  spiralAngle?: number;
  hobspeedRoughing?: number;
  hobspeedFinishing?: number;
  gearAddendum?: number;
  toothDepth?: number;
  depthOfCutRoughing?: number;
  depthOfCutFinishing?: number;
  hobApproachRoughing?: number;
  hobApproachFinishing?: number;
  hobOverrun?: number;
  totalHobRoughing?: number;
  totalHobFinishing?: number;
  operationTypeId?: number;
  operationName?: string;
  featureId?: string;
  blankArea?: number;
  formingForce?: number;
  noOfNodePoints?: number;
  harnessRequirement?: number;
  typeOfSplice?: number;
  maxLength?: number;
  minLength?: number;
  additionalLengthArray?: string;
  isWorkpieceOuterDiaDirty?: boolean;
  cuttingSpeed?: number;
  feed?: number;
  depthOfCut?: number;
  subProcessTypeID?: number;
  surfaceArea?: number;
  volume?: number;
}